Discord auth and user creation
We create a User model which will be the basis for auth and stuff. We
hit the discord auth endpoint and get back all the user data we want.
Then we go through and log the user in if they already have an account
or create a new one if they don't.

Created our auth methods in `ApplicationController` - authenticate and
current_user are the big ones

Placeholder dashboard with random user data.