Display registrations on tournament page
Displayed the registered players / number of them on the tournament
page. Show empty if there are empty slots. Show registered in the top if
the current_user is registered with a button to drop.

Added a slot number to the tournament table - represents the number of
players the tournament can support.

Added ratings to users.