Show match slip if in match
If the `current_user` is currently in a match there should be a match
slip at the top of every page. That match slip should list the current
match they're in with a form to set the score of the round.

Added the concept of an active_player to the user to make it easier to
get the `@tournament` and `@round` - also adding a status to rounds to
make it clear which to query for. when a score is submitted the round
will switch to completed.

We're displaying the pairings now on the tournament show page.